add confirmation to delete reervation and check the functionality

+++ app/reservations/ReservationClient.tsx
@@ -19,30 +19,30 @@ export default function ReservationClient({reservations, currentUser}: TripsClie
     const [deletingId, setDeletingId] = useState("")
 
     const onCancel = useCallback((id: string) => {
-        setDeletingId(id)
 
-        axios.delete(`/api/reservations/${id}`).then(() => {
-            toast.success("Reservation cancelled");
-            router.refresh()
+        let result = confirm(" Are you sure you want to remove the reservation ? ");
+        if (result) {
+            setDeletingId(id)
+            axios.delete(`/api/reservations/${id}`).then(() => {
+                toast.success("Reservation cancelled");
+                router.refresh()
 
-        }).catch((error) => {
+            }).catch((error) => {
+                toast.error(error?.response?.data?.error)
+            }).finally(() =>
 
-            toast.error(error?.response?.data?.error)
-        }).finally(() =>
-
-            setDeletingId("")
-        )
-
+                setDeletingId("")
+            )
+        }
 
     }, [router])
 
     return (
         <Container>
-            <Heading title="Reservations!" subtitle="Booking on your properties!"/>
+            <Heading title="Reservations!" subTitle="Booking on your properties!"/>
             <div
                 className="mt-10 grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 lg:grid-cols-4 xl:grid-cols-5 2xl:grid-cols-6 gap-8">
                 {reservations.map((reservation) => (
-
                     <ListingCard
                         key={reservation.id}
                         data={reservation.listing}
