+++ app/trips/TripsClient.tsx
@@ -9,7 +9,6 @@ import {toast} from "react-hot-toast";
 import ListingCard from "@/app/components/listings/ListingCard";
 
 interface TripsClientProps {
-
     reservations: SafeReservations[];
     currentUser?: safeUser | undefined;
 }
@@ -19,26 +18,28 @@ export default function TripsClient({reservations, currentUser}: TripsClientProp
     const [deletingId, setDeletingId] = useState("")
 
     const onCancel = useCallback((id: string) => {
-        setDeletingId(id)
+        let result = confirm("Are you sure you want to remove the reservation ?");
+        if (result) {
 
-        axios.delete(`/api/reservations/${id}`).then(() => {
-            toast.success("Reservation cancelled");
-            router.refresh()
+            setDeletingId(id)
 
-        }).catch((error) => {
+            axios.delete(`/api/reservations/${id}`).then(() => {
+                toast.success("Reservation cancelled");
+                router.refresh()
+            }).catch((error) => {
+                toast.error(error?.response?.data?.error)
+            }).finally(() =>
+                setDeletingId("")
+            )
+        }
 
-            toast.error(error?.response?.data?.error)
-        }).finally(() =>
-
-            setDeletingId("")
-        )
 
 
     }, [router])
 
     return (
         <Container>
-            <Heading title="Trips!" subtitle="where you have been and where you're going!"/>
+            <Heading title="Trips!" subTitle="where you have been and where you're going!"/>
             <div
                 className="mt-10 grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 lg:grid-cols-4 xl:grid-cols-5 2xl:grid-cols-6 gap-8">
                 {reservations.map((reservation) => (
